craft: add crypt by file,README
