+++ craft/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "craft"
+name = "git-craft"
 version = "0.1.0"
 edition = "2021"
+
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -10,3 +11,5 @@ anyhow = "1.0.69"
 pgp = "0.9.0"
 rand = "0.8.5"
 smallvec = "1.10.0"
+clap = { version = "4.3.0", features = ["derive"] }
+git = {path = "../git"}
