+++ craft/src/message.txt
@@ -1,0 +1,3 @@
+hello,world
+$#$&##@(*
+绍兴花雕二年陈
